+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.26
 
 - Improved error messages when writing external files. Mirror of Leo's #4732
+- Improved documentation and help commands for the 'Abbreviations' features.
 
 ## 1.0.25