+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.0.27
+
+- Fixed broken links in documentation
+
 ## 1.0.26
 
 - Added documentation with docusaurus at /docs/.